refactor: move Metadata from store/mod.rs to metadata.rs
